DSSI: Getting system to rebuild
- Minimal changes made to get dssi to compile and run again
- Resampling should be looked into
- Additional change-sets may appear from zynaddsubfx:dssi (experimental) or
  yoshimi git repos

+++ src/Misc/Master.h
@@ -100,7 +100,7 @@ class Master
         /**Audio Output*/
         void AudioOut(float *outl, float *outr);
         /**Audio Output (for callback mode). This allows the program to be controled by an external program*/
-        void GetAudioOutSamples(int nsamples,
+        void GetAudioOutSamples(size_t nsamples,
                                 int samplerate,
                                 float *outl,
                                 float *outr);
